// src/app/calculator/entities/interfaces/experience.interface.ts
export interface IExperience {
  woodcutting?: number;
  mining?: number;
  fishing?: number;
  alchemy?: number;
  smelting?: number;
  cooking?: number;
  forge?: number;
  shadowMastery?: number;
  bartering?: number;
  huntingMastery?: number;
  yuleMastery?: number;
  springtideMastery?: number;
  combat?: number;
  petMastery?: number;
  guildMastery?: number;
  lunarMastery?: number;
  strength?: number;
  defence?: number;
  speed?: number;
  dexterity?: number;
}